+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added endpoint `get_flow_temperature()` to the heat circuit
+- Added endpoint `get_return_flow_temperature()` to the heat circuit
+
 ## [2.0.0] - 2025-09-30
 
 ### Added